requirements.txt
mcp
groq
chromadb
sentence-transformers
python-dotenv